requirements.txt
discord.py[voice]>=2.4.0
python-dotenv>=1.0.0
PyNaCl>=1.5.0
groq>=0.9.0
edge-tts>=6.1.9
supabase>=2.4.0